+++ src/outputs/cluster_manifests/application-manifest.ts
