+++ src/outputs/cluster_manifests/cert-manager/production-cluster-issuer.ts
