+++ src/outputs/cluster_manifests/cert-manager/self-signed/dev-cluster-issuer.ts
