+++ src/outputs/cluster_manifests/core-applications/cert-manager.application.yaml.ts
