+++ src/outputs/cluster_manifests/core-applications/external-dns.application.yaml.ts
