+++ src/outputs/cluster_manifests/core-applications/kube-prometheus-stack.application.yaml.ts
