+++ src/outputs/cluster_manifests/core-applications/loki.application.yaml.ts
