+++ src/outputs/cluster_manifests/core-applications/private-nginx.application.yaml.ts
