+++ src/outputs/cluster_manifests/core-applications/promtail.application.yaml.ts
