+++ src/outputs/cluster_manifests/core-applications/public-nginx.application.yaml.ts
