+++ src/outputs/cluster_manifests/core-applications/reloader.application.yaml.ts
