+++ src/outputs/cluster_manifests/functions/fns-ingress.ts
@@ -1,4 +1,4 @@
-import getPublicIngressManifest from "src/outputs/public-ingress-manifest.ts";
+import getPublicIngressManifest from "src/outputs/cluster_manifests/ingress/public-ingress-manifest.ts";
 
 export function getFunctionsIngressManifest(hostname: string): string {
   return getPublicIngressManifest("fns-ingress", {
