+++ src/outputs/cluster_manifests/functions/fns-namespace.ts
