+++ src/outputs/cluster_manifests/functions/fns-pull-secret.ts
