+++ src/outputs/cluster_manifests/functions/fns-service.ts
