+++ src/outputs/cluster_manifests/ingress/argo-ingress.yaml.ts
@@ -1,5 +1,4 @@
-//
-import getPublicIngressManifest from "src/outputs/public-ingress-manifest.ts";
+import getPublicIngressManifest from "src/outputs/cluster_manifests/ingress/public-ingress-manifest.ts";
 
 export function getArgoIngressManifest(hostname: string): string {
   return getPublicIngressManifest("argocd-ingress", {
