+++ src/outputs/cluster_manifests/ingress/microk8s/ingress-daemonset.ts
