+++ src/outputs/cluster_manifests/ingress/microk8s/ingress-tcp-services-configmap.ts
