+++ src/outputs/cluster_manifests/ingress/public-ingress-manifest.ts
