+++ src/outputs/cluster_manifests/sealed-secret-manifest.ts
